+++ src/row.rs
@@ -101,7 +101,7 @@ impl Row {
     /// let conn = Connection::connect("scott", "tiger", "", &[])?;
     /// let mut stmt = conn.prepare("select empno, ename from emp", &[])?;
     ///
-    /// for result in stmt.query(&[])? {
+    /// for result in &stmt.query(&[])? {
     ///     let row = result?;
     ///     // Gets a row as `(i32, String)`.
     ///     let (empno, ename) = row.get_as::<(i32, String)>()?;
@@ -153,7 +153,7 @@ impl<'a, T> ResultSet<'a, T> where T: RowValue {
     }
 }
 
-impl<'stmt, T> Iterator for ResultSet<'stmt, T> where T: RowValue {
+impl<'a, 'stmt, T> Iterator for &'a ResultSet<'stmt, T> where T: RowValue {
     type Item = Result<<T>::Item>;
 
     fn next(&mut self) -> Option<Self::Item> {
@@ -217,7 +217,7 @@ impl<'stmt, T> Iterator for ResultSet<'stmt, T> where T: RowValue {
 /// let mut stmt = conn.prepare("select * from emp", &[])?;
 ///
 /// // Gets rows as Emp
-/// for result in stmt.query_as::<Emp>(&[])? {
+/// for result in &stmt.query_as::<Emp>(&[])? {
 ///     let emp = result?;
 ///     println!("{},{}", emp.empno, emp.ename);
 /// }